Video: Fix the Auto internal resolution scaling for unusual aspect ratios
The automatic resolution scaling didn’t work properly for unusual aspect ratios (like 32:9). It was including the parts of the screen that would show black bars in its calculations, which led to incorrect scaling and resolution matching. This change fixes that by only focusing on the actual part of the screen where the image will be displayed.

The original code was modified, so we can use the non-integer scaling factor values in MMJR2.