New versioning system
The old versioning system was partly inherited from MMJR2 and always felt somewhat arbitrary. The new system reflects the codebase of the official Dolphin project that MMJR2-VBI is based on.

For development builds, I will use the date of the latest MMJR2-VBI release: DEV-RR.MM.DD.

For release and beta builds, I will use the date and upstream version to mark the codebase it is built upon. For example, for Dolphin 5.0-20594 from December 10th, 2023, the version would be U23.12.10-20594.