Set OSD color to yellow in deafult AMOLED theme
Purple is hard to read on dark backgrounds, so this changes the default OSD color back to the more readable yellow used in official Dolphin.

The code in OnScreenDisplay.h now matches official Dolphin. Color values are still overwritten by the AndroidTheme system. The colors for the OSD are defined in AndroidTheme.h and AndroidTheme.cpp, where the defaults are now set to yellow.